requirements.txt
# ----- web/infra -----
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1

# ----- DL stack (CUDA 11.8 channel에서 받음) -----
# torch/torchvision/torchaudio는 pip --index-url 로 받되, version은 상호 호환 세트로 고정
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

# ----- ml/etl -----
scikit-learn==1.5.1
pillow==10.4.0
albumentations==1.4.7

mlflow==2.14.1

dvc==3.50.0
dvc-s3==3.2.0

prometheus-client==0.20.0

opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-sdk==1.26.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-semantic-conventions==0.47b0

timm>=1.0.9
